Create first_hash_map.py
My first creation of a hash map. A hash map makes it easier to find data in a dictionary by compiling the data points into an array and assigning a 'hash' value to a key instead of itterating through an entire list. There are many ways to deal with a key giving the same index value. This hash map simply moves on to a later index value by a degree of 1. Other examples include: itterating by another value or power of 2, 3, etc or creating another array inside each hash map array value